Move code for computing Intersection Observations from document to IntersectionObserver https://bugs.webkit.org/show_bug.cgi?id=259173 rdar://112176104
Reviewed by Cameron McCormack.

Move most the Intersetion Observer-related code in Document into IntersectionObserver.cpp
for better encapsulation.

The only behavior change is to keep RefPtr to the IntersectionObserver while running the code.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::updateIntersectionObservations):
(WebCore::expandRootBoundsWithRootMargin): Deleted.
(WebCore::computeClippedRectInRootContentsSpace): Deleted.
(WebCore::computeIntersectionState): Deleted.
* Source/WebCore/html/LazyLoadFrameObserver.cpp:
* Source/WebCore/page/IntersectionObserver.cpp:
(WebCore::expandRootBoundsWithRootMargin):
(WebCore::computeClippedRectInRootContentsSpace):
(WebCore::computeIntersectionState):
(WebCore::IntersectionObserver::updateObservations):
* Source/WebCore/page/IntersectionObserver.h:

Canonical link: https://commits.webkit.org/266034@main
